# openapi/test-runtime.openapi.yaml
openapi: 3.0.3
info:
  title: Keyhole Test Runtime API
  description: Public test runtime for the Keyhole developer ecosystem.
  version: 0.1.0
paths:
...
  /realize:
    post:
      summary: Realize a candidate digest
      operationId: realize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RealizationRequest"
      responses:
        "200":
          description: Realization receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealizationReceipt"
        "422":
          description: Validation error
components:
  schemas:
...
    RealizationRequest:
      type: object
      required: [candidate_digest]
      properties:
        candidate_digest:
          type: string
          example: sha256:abc123
        payload:
          type: object
          nullable: true
    RealizationReceipt:
      type: object
      required: [digest, status, message, realized_at]
      properties:
        digest:
          type: string
          example: sha256:abc123
        status:
          type: string
          enum: [ACCEPT, ALREADY_REALIZED]
          example: ACCEPT
        message:
          type: string
          example: Digest realized successfully.
        realized_at:
          type: string
          format: date-time
          example: "2026-03-06T12:01:00+00:00"